Remove todo item for v0.2.0
This item would have added a bloom filter to avoid the bot from repeating itself when making new messages. I have decided to keep this behavior in for now, because the chances of it sending the same message twice in a row is slim, and it would be difficult to determine when an infinite loop would occur as a result of this.